backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-filter==23.5
djoser==2.3.3
djangorestframework-simplejwt==5.5.1
Pillow==12.0.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
python-dotenv==1.0.1
django-cors-headers==4.4.0
setuptools
webcolors==1.11.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0.1
